refactor(analytics): extract ChampionBuildComputeService from the god-file (P10.1)
Continues the P10.1 decomposition by lifting the builds domain out of
ChampionAnalyticsComputeService (+ its .Stats.cs partial) into a focused
ChampionBuildComputeService, mirroring the win-rate extraction (#100).

- New IChampionBuildComputeService + ChampionBuildComputeService own
  ComputeBuildsAsync (and its 3 build-only consts) + ComputeBuildsFromStatsAsync,
  moved verbatim. The new service takes the logger (build compute logs item-
  metadata-coverage fallbacks); win-rate/tier-list, pro builds/playrate, and
  matchups stay on the original service.
- Reuses the shared AnalyticsSampleThreshold helper (no duplicated sample rule)
  and the already-extracted ChampionBuildPathBuilder.
- Rewired the orchestrator (ChampionAnalyticsService) and the
  PrecomputedAnalyticsRefresher to inject + route to the build service; DI
  registers it alongside the win-rate service.
- Removed the build signatures from IChampionAnalyticsComputeService and fixed
  the stale .Stats.cs class doc.

Behavior-preserving, gated by the analytics tests (raw + raw-vs-stats build
equivalence): Service.Core 183/183, WebAPI 40/40. No asserted values changed.
