+++ Transcendence.Service.Core/Services/Analytics/Interfaces/IChampionAnalyticsComputeService.cs
@@ -8,32 +8,6 @@ namespace Transcendence.Service.Core.Services.Analytics.Interfaces;
 /// </summary>
 public interface IChampionAnalyticsComputeService
 {
-    /// <summary>
-    /// Computes top 3 builds for a champion with items and runes bundled.
-    /// Core items (70%+ appearance) distinguished from situational.
-    /// Uses only completed, build-impact items for build quality calculations.
-    /// </summary>
-    Task<ChampionBuildsResponse> ComputeBuildsAsync(
-        int championId,
-        string role,
-        string? rankTier,
-        string? region,
-        string patch,
-        CancellationToken ct);
-
-    /// <summary>
-    /// Builds served from the durable <c>ChampionBuildSnapshot</c> (the persisted live response for the
-    /// common all-region scopes), falling back to <see cref="ComputeBuildsAsync"/> for a specific tier/region
-    /// or a patch without a snapshot yet. The stored value is the live compute's own output, so it's identical.
-    /// </summary>
-    Task<ChampionBuildsResponse> ComputeBuildsFromStatsAsync(
-        int championId,
-        string role,
-        string? rankTier,
-        string? region,
-        string patch,
-        CancellationToken ct);
-
     Task<ChampionProBuildsResponse> ComputeProBuildsAsync(
         int championId,
         string? region,
